Rename TC to PR number
