+++ apps/src/tests/issue-tests/index.ts
@@ -183,8 +183,8 @@ export { default as Test3611 } from './Test3611';
 export { default as Test3617 } from './Test3617';
 export { default as Test3760 } from './Test3760';
 export { default as Test3770 } from './Test3770';
+export { default as Test3793 } from './Test3793';
 export { default as Test3816 } from './Test3816';
-export { default as TestXXXX } from './TestXXXX';
 export { default as TestScreenAnimation } from './TestScreenAnimation';
 // The following test was meant to demo the "go back" gesture using Reanimated
 // but the associated PR in react-navigation is currently put on hold
